requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.95.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
click==8.3.2
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.135.3
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
jiter==0.14.0
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-instrumentation==0.62b0
opentelemetry-instrumentation-asgi==0.62b0
opentelemetry-instrumentation-fastapi==0.62b0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opentelemetry-util-http==0.62b0
packaging==26.1
prometheus_client==0.25.0
protobuf==6.33.6
pydantic==2.13.1
pydantic_core==2.46.1
python-dotenv==1.2.2
sniffio==1.3.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
wrapt==2.1.2
zipp==3.23.1
